Move Toggle into separate file
